tox.ini: use `source_pkgs` in [coverage:run] section
Since the package name "tox_pipenv" and the
source directory are the same, passing
`--cov tox_pipenv` results in pytest-cov taking
the arg as a filesystem path, rather than a package
name. Using a filesystem path here is inconvenient
because the package gets installed into a virtualenv
and doesn't run from the repo root.

The previous workaround was to explicitly pass
the {envsitepackagesdir} to `--cov`, but _this_
technique didn't work for subprocesses that
change the working directory (see pytest-cov#465).

This workaround resolves the subprocess coverage
issue, but retains the paths mapping so that the
output references the source files in the top-level
tox_pipenv directory.